Large commit (read description)
In this commit added:

1) Completed socket logic between client and server

2) Completed dialog layout with next features: editing/deleting/copying message, deleting entrie chat, blocking user, supports emoji

3) Font fix for phone

4) Some UI changes

5) Added Docker supports

+++ backend/src/services/ChatService.ts
@@ -1,5 +1,6 @@
 import { prisma } from "../utils/handlers";
 import { encrypt, decrypt } from "../utils/crypt";
+import { getChatIdByUsersId } from "../utils/index";
 
 export default class ChatService {
   static async getChatsByUserId(userId: string) {
@@ -54,20 +55,78 @@ export default class ChatService {
       );
   }
 
-  static async getMessagesByChatId(chatId: string) {
-    const messages = await prisma.message.findMany({ where: { chatId } });
-    return messages.map((message) => ({
+  static async setIsReadedMessagesByChatIdAndRecepientId(
+    chatId: string,
+    userId: string
+  ) {
+    const messages = await prisma.message.findMany({
+      where: { chatId, userId, isReaded: false },
+      select: { id: true },
+    });
+
+    await prisma.message.updateMany({
+      where: { chatId, userId, isReaded: false },
+      data: { isReaded: true },
+    });
+
+    return messages.map((message) => message.id);
+  }
+
+  static async getMessagesByChatId(
+    page: number,
+    limit: number,
+    chatId: string
+  ) {
+    const messages = await prisma.message.findMany({
+      where: { chatId },
+      orderBy: { createdAt: "desc" },
+      select: {
+        text: true,
+        id: true,
+        createdAt: true,
+        isReaded: true,
+        isEdited: true,
+        userId: true,
+      },
+      skip: (page - 1) * limit,
+      take: limit,
+    });
+
+    const decryptedMessages = messages.map((message) => ({
       ...message,
       text: decrypt(message.text),
     }));
+
+    const totalMessagesCount = await prisma.message.count({
+      where: { chatId },
+    });
+
+    return {
+      messages: decryptedMessages,
+      hasMore: page * limit < totalMessagesCount,
+    };
   }
 
   static async deleteChatById(id: string) {
+    const chat = prisma.chat.findUnique({ where: { id } });
+    if (!chat) throw new Error("No chat found");
+
     await prisma.chat.delete({ where: { id } });
   }
 
   static async deleteMessageById(id: string) {
+    const message = await prisma.message.findUnique({ where: { id } });
+    if (!message) throw new Error("No message found");
+
     await prisma.message.delete({ where: { id } });
+
+    const messagesCount = await prisma.message.count({
+      where: { chatId: message.chatId },
+    });
+
+    if (!messagesCount) {
+      await this.deleteChatById(message.chatId);
+    }
   }
 
   static async createChat(userId_1: string, userId_2: string) {
@@ -78,14 +137,24 @@ export default class ChatService {
       throw new Error("No users provided");
     }
 
-    const chat = await prisma.chat.create({});
+    return await prisma.$transaction(async (tx) => {
+      const chat = await tx.chat.create({
+        data: { id: getChatIdByUsersId(userId_1, userId_2) },
+      });
 
-    await prisma.chatParticipant.createMany({
-      data: [
-        { userId: userId_1, chatId: chat.id },
-        { userId: userId_2, chatId: chat.id },
-      ],
+      await tx.chatParticipant.createMany({
+        data: [
+          { userId: userId_1, chatId: chat.id },
+          { userId: userId_2, chatId: chat.id },
+        ],
+      });
+
+      return chat;
     });
+  }
+
+  static async isChatExist(chatId: string) {
+    return await prisma.chat.findUnique({ where: { id: chatId } });
   }
 
   static async createMessage(userId: string, chatId: string, text: string) {
@@ -94,13 +163,66 @@ export default class ChatService {
       throw new Error("Chat not found");
     }
 
-    return await prisma.message.create({
+    const message = await prisma.message.create({
       data: {
         userId,
         chatId,
         text: encrypt(text),
       },
+      select: {
+        text: true,
+        id: true,
+        createdAt: true,
+        isEdited: true,
+        isReaded: true,
+        userId: true,
+      },
     });
+
+    return { ...message, text: decrypt(message.text) };
+  }
+
+  static async setIsReadedMessageById(id: string, isReaded: boolean) {
+    await prisma.message.update({ where: { id }, data: { isReaded } });
+    const message = await prisma.message.findUnique({
+      where: { id },
+      select: {
+        text: true,
+        id: true,
+        createdAt: true,
+        isEdited: true,
+        isReaded: true,
+        userId: true,
+      },
+    });
+
+    if (!message) {
+      throw new Error("No message found");
+    }
+
+    return { ...message, text: decrypt(message.text) };
+  }
+
+  static async updateMessageById(id: string, text: string) {
+    const message = await prisma.message.update({
+      where: { id },
+      data: { text: encrypt(text), isEdited: true, isReaded: false },
+      select: {
+        text: true,
+        id: true,
+        createdAt: true,
+        isEdited: true,
+        isReaded: true,
+        userId: true,
+      },
+    });
+
+    return { ...message, text: decrypt(message.text) };
+  }
+
+  static async getChatParticipantIdsByUserId(id: string) {
+    const chats = await this.getChatsByUserId(id);
+    return chats.map((chat) => chat.userId);
   }
 
   static async deleteAllChatsByUserId(userId: string) {
@@ -116,5 +238,20 @@ export default class ChatService {
     await prisma.chat.deleteMany({
       where: { id: { in: chatIds } },
     });
+  }
+
+  static async isUserInChat(userId: string, chatId: string) {
+    const inChat = await prisma.chatParticipant.findUnique({
+      where: {
+        chatId_userId: {
+          chatId,
+          userId,
+        },
+      },
+    });
+
+    if (!inChat) {
+      throw new Error("User is not a participant in the chat");
+    }
   }
 }
